UI Framework
Set up a new UI system for mini golf, based on the old one. Keeps track of what hole you're on, how many shots you've taken this hole, and what your current ability is.

Also, when the hole is over, a scoreboard now displays. But it isn't working just yet.