Working Scoreboard
The scoreboard now works. At the end of each hole, you're told what your score for the hole was. If you also set up the scene to have a holeNumber, it updates the right entry in the actual scorecard as well. See the scene "Hole1" for an example of this.