Ability Overhaul
THIS CODE WILL BREAK LEVELS! PLEASE READ!

All scripts related to the Ability system were duplicated and rewritten, to make a reworked system. I've applied this system to other scripts, but not yet prefab objects, such as the player's golf ball. You will get script errors when you try to load in old scenes. I'm going to try and fix this with prefab updates ASAP.

I'll go over a summary of the abilities in discord- but if you want to see everything in action for yourself, visit the "AllAbilitiesTest" scene. The only abilities not yet in a working state are Labyrinth and Zero Gravity, but I'm sure they'll come along soon!