Labyrinth (prototype)
Working prototype for the labyrinth ability! Try it out by adding it to your scene in Prefabs. All you need to do is set "courseContainer" to the object you want to rotate, and attach a LabyrinthRotate script to that object. If you want to rotate the entire map, put everything in an empty object.

I'll add a few more improvements later, such as resetting the rotation upon death and disabling ability pickup while you are labyrinth-ing.