Shot stats UI framework
The bottom portion of "controls" will soon give you precise info about your shot, including angle, loft and power (as well as a reminder of how to shoot the darned thing, by pressing space or LMB.) It's not finished yet, but you can see a test of the direction arrow seems to be working. So that's a positive.